requirements.txt
Django>=3.2.4,<3.3
djangorestframework>=3.12.4,<3.13
psycopg2>=2.8.6,<2.9
drf-spectacular>=0.15.1,<0.16
django-mptt==0.13.4
model_bakery>=1.8.0,<1.9.0
factory_boy>=3.2.0,<3.2.1
Pillow>=9.2,<9.3
django-paypal>=1.1.2,<2.0